requirements.txt
Flask
Flask-HTTPAuth
pandas
numpy
scikit-learn
joblib
load_dotenv